+++ src/Client/AndroidBall.h
@@ -51,5 +51,10 @@ class AndroidBall : public CCLayer
 
 class $modify (QOLModTouchDispatcher, CCTouchDispatcher)
 {
+    static void onModify(auto& self)
+    {
+        self.setHookPriority("cocos2d::CCTouchDispatcher::touches", -99999999);
+    }
+
     void touches(CCSet* touches, CCEvent* event, unsigned int type);
 };