+++ changelog.md
@@ -1,9 +1,12 @@
 # 1.6.7
 
+- Fixed Speedhack not working with cbf enabled
 - Added **Toolbox Button Bypass**
 - Added **Scale Text Input**
 - Added **Free Scroll**
 - Added **No Trail Blending**
+- Added **Percentage Level Font**
+- Added **Label Font Customizer**
 
 # 1.6.6